Create histogram.rb
In this project, we'll build a program that takes a user's input, then builds a hash from that input. Each key in the hash will be a word from the user; each value will be the number of times that word occurs. A visual representation of data like this is called a histogram.